+++ pytket/extensions/qiskit/backends/ibm.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
 import itertools
 import logging
 from ast import literal_eval
@@ -23,16 +24,17 @@ from typing import (
     List,
     Optional,
     Dict,
-    Any,
     Sequence,
     TYPE_CHECKING,
     Tuple,
     Union,
+    Set,
+    Any,
 )
 from warnings import warn
 
-import qiskit  # type: ignore
-from qiskit import IBMQ
+from qiskit_ibm_provider import IBMProvider  # type: ignore
+from qiskit_ibm_provider.exceptions import IBMProviderError  # type: ignore
 from qiskit.primitives import SamplerResult  # type: ignore
 
 
@@ -72,7 +74,9 @@ from pytket.passes import (  # type: ignore
     CliffordSimp,
     SimplifyInitial,
     NaivePlacementPass,
+    RebaseCustom,
 )
+from pytket.passes._decompositions import _TK1_to_X_SX_Rz
 from pytket.predicates import (  # type: ignore
     NoMidMeasurePredicate,
     NoSymbolsPredicate,
@@ -92,10 +96,7 @@ from .ibm_utils import _STATUS_MAP, _batch_circuits
 from .config import QiskitConfig
 
 if TYPE_CHECKING:
-    from qiskit.providers.ibmq import (  # type: ignore
-        IBMQBackend as _QiskIBMQBackend,
-        AccountProvider,
-    )
+    from qiskit_ibm_provider.ibm_backend import IBMBackend as _QiskIBMBackend  # type: ignore
 
 _DEBUG_HANDLE_PREFIX = "_MACHINE_DEBUG_"
 
@@ -110,7 +111,7 @@ def _gen_debug_results(n_qubits: int, shots: int, index: int) -> SamplerResult:
     )
 
 
-class NoIBMQAccountError(Exception):
+class NoIBMQCredentialsError(Exception):
     """Raised when there is no IBMQ account available for the backend"""
 
     def __init__(self) -> None:
@@ -124,17 +125,46 @@ def _save_ibmq_auth(qiskit_config: Optional[QiskitConfig]) -> None:
     token = None
     if qiskit_config is not None:
         token = qiskit_config.ibmq_api_token
-    if not IBMQ.active_account():
-        if IBMQ.stored_account():
-            IBMQ.load_account()
+    if token is None and os.getenv("PYTKET_REMOTE_QISKIT_TOKEN") is not None:
+        token = os.getenv("PYTKET_REMOTE_QISKIT_TOKEN")
+    try:
+        if token is not None:
+            IBMProvider.save_account(token, overwrite=True)
+            IBMProvider()
         else:
-            if token is not None:
-                IBMQ.save_account(token)
-            else:
-                raise NoIBMQAccountError()
+            IBMProvider()
+    except:
+        if token is not None:
+            IBMProvider.save_account(token, overwrite=True)
+            IBMProvider()
+        else:
+            raise NoIBMQCredentialsError()
     if not QiskitRuntimeService.saved_accounts():
         if token is not None:
             QiskitRuntimeService.save_account(channel="ibm_quantum", token=token)
+
+
+# Variables for defining a rebase to the {X, SX, Rz, ECR} gateset
+# See https://github.com/CQCL/pytket-qiskit/issues/112
+_cx_replacement_with_ecr = (
+    Circuit(2).X(0).SX(1).Rz(-0.5, 0).add_gate(OpType.ECR, [0, 1])
+)
+_tk1_replacement_function = _TK1_to_X_SX_Rz
+
+ecr_rebase = RebaseCustom(
+    gateset={OpType.X, OpType.SX, OpType.Rz, OpType.ECR},
+    cx_replacement=_cx_replacement_with_ecr,
+    tk1_replacement=_tk1_replacement_function,
+)
+
+
+def _get_primitive_gates(gateset: Set[OpType]) -> Set[OpType]:
+    if gateset >= {OpType.X, OpType.SX, OpType.Rz, OpType.CX}:
+        return {OpType.X, OpType.SX, OpType.Rz, OpType.CX}
+    elif gateset >= {OpType.X, OpType.SX, OpType.Rz, OpType.ECR}:
+        return {OpType.X, OpType.SX, OpType.Rz, OpType.ECR}
+    else:
+        return gateset
 
 
 class IBMQBackend(Backend):
@@ -146,11 +176,9 @@ class IBMQBackend(Backend):
     def __init__(
         self,
         backend_name: str,
-        hub: Optional[str] = None,
-        group: Optional[str] = None,
-        project: Optional[str] = None,
+        instance: Optional[str] = None,
         monitor: bool = True,
-        account_provider: Optional["AccountProvider"] = None,
+        provider: Optional["IBMProvider"] = None,
         token: Optional[str] = None,
     ):
         """A backend for running circuits on remote IBMQ devices.
@@ -159,35 +187,26 @@ class IBMQBackend(Backend):
         using :py:meth:`pytket.extensions.qiskit.set_ibmq_config`.
         This function can also be used to set the IBMQ API token.
 
-        :param backend_name: Name of the IBMQ device, e.g. `ibmqx4`,
-         `ibmq_16_melbourne`.
+        :param backend_name: Name of the IBMQ device, e.g. `ibmq_16_melbourne`.
         :type backend_name: str
-        :param hub: Name of the IBMQ hub to use for the provider.
-         If None, just uses the first hub found. Defaults to None.
-        :type hub: Optional[str], optional
-        :param group: Name of the IBMQ group to use for the provider. Defaults to None.
-        :type group: Optional[str], optional
-        :param project: Name of the IBMQ project to use for the provider.
-         Defaults to None.
-        :type project: Optional[str], optional
+        :param instance: String containing information about the hub/group/project.
+        :type instance: str, optional
         :param monitor: Use the IBM job monitor. Defaults to True.
         :type monitor: bool, optional
         :raises ValueError: If no IBMQ account is loaded and none exists on the disk.
-        :param account_provider: An AccountProvider returned from IBMQ.enable_account.
-         Used to pass credentials in if not configured on local machine (as well as hub,
-         group and project). Defaults to None.
-        :type account_provider: Optional[AccountProvider]
+        :param provider: An IBMProvider
+        :type provider: Optional[IBMProvider]
         :param token: Authentication token to use the `QiskitRuntimeService`.
         :type token: Optional[str]
         """
         super().__init__()
         self._pytket_config = QiskitConfig.from_default_config_file()
         self._provider = (
-            self._get_provider(hub, group, project, self._pytket_config)
-            if account_provider is None
-            else account_provider
+            self._get_provider(instance=instance, qiskit_config=self._pytket_config)
+            if provider is None
+            else provider
         )
-        self._backend: "_QiskIBMQBackend" = self._provider.get_backend(backend_name)
+        self._backend: "_QiskIBMBackend" = self._provider.get_backend(backend_name)  # type: ignore
         config = self._backend.configuration()
         self._max_per_job = getattr(config, "max_experiments", 1)
 
@@ -197,7 +216,9 @@ class IBMQBackend(Backend):
         self._service = QiskitRuntimeService(channel="ibm_quantum", token=token)
         self._session = Session(service=self._service, backend=backend_name)
 
-        self._standard_gateset = gate_set >= {OpType.X, OpType.SX, OpType.Rz, OpType.CX}
+        self._primitive_gates = _get_primitive_gates(gate_set)
+
+        self._supports_rz = OpType.Rz in self._primitive_gates
 
         self._monitor = monitor
 
@@ -208,33 +229,16 @@ class IBMQBackend(Backend):
 
     @staticmethod
     def _get_provider(
-        hub: Optional[str],
-        group: Optional[str],
-        project: Optional[str],
+        instance: Optional[str],
         qiskit_config: Optional[QiskitConfig],
-    ) -> "AccountProvider":
+    ) -> "IBMProvider":
         _save_ibmq_auth(qiskit_config)
-        provider_kwargs: Dict[str, Optional[str]] = {}
-        if hub:
-            provider_kwargs["hub"] = hub
-        else:
-            provider_kwargs["hub"] = qiskit_config.hub if qiskit_config else None
-        if group:
-            provider_kwargs["group"] = group
-        else:
-            provider_kwargs["group"] = qiskit_config.group if qiskit_config else None
-        if project:
-            provider_kwargs["project"] = project
-        else:
-            provider_kwargs["project"] = (
-                qiskit_config.project if qiskit_config else None
-            )
         try:
-            if any(x is not None for x in provider_kwargs.values()):
-                provider = IBMQ.get_provider(**provider_kwargs)
+            if instance is not None:
+                provider = IBMProvider(instance=instance)
             else:
-                provider = IBMQ.providers()[0]
-        except qiskit.providers.ibmq.exceptions.IBMQProviderError as err:
+                provider = IBMProvider()
+        except IBMProviderError as err:
             logging.warn(
                 (
                     "Provider was not specified enough, specify hub,"
@@ -250,7 +254,7 @@ class IBMQBackend(Backend):
         return self._backend_info
 
     @classmethod
-    def _get_backend_info(cls, backend: "_QiskIBMQBackend") -> BackendInfo:
+    def _get_backend_info(cls, backend: "_QiskIBMBackend") -> BackendInfo:
         config = backend.configuration()
         characterisation = process_characterisation(backend)
         averaged_errors = get_avg_characterisation(characterisation)
@@ -284,7 +288,7 @@ class IBMQBackend(Backend):
         gate_set = _tk_gate_set(backend)
         backend_info = BackendInfo(
             cls.__name__,
-            backend.name(),
+            backend.name,
             __extension_version__,
             arch,
             gate_set.union(
@@ -310,11 +314,15 @@ class IBMQBackend(Backend):
 
     @classmethod
     def available_devices(cls, **kwargs: Any) -> List[BackendInfo]:
-        provider: Optional["AccountProvider"] = kwargs.get("account_provider")
+        provider: Optional["IBMProvider"] = kwargs.get("provider")
         if provider is None:
-            provider = cls._get_provider(
-                kwargs.get("hub"), kwargs.get("group"), kwargs.get("project"), None
-            )
+            if kwargs.get("instance") is not None:
+                provider = cls._get_provider(
+                    instance=kwargs.get("instance"), qiskit_config=None
+                )
+            else:
+                provider = IBMProvider()
+
         backend_info_list = [
             cls._get_backend_info(backend) for backend in provider.backends()
         ]
@@ -389,7 +397,10 @@ class IBMQBackend(Backend):
         # https://cqcl.github.io/pytket-qiskit/api/index.html#default-compilation
         # Edit this docs source file -> pytket-qiskit/docs/intro.txt
         if optimisation_level == 0:
-            if self._standard_gateset:
+            if self._supports_rz:
+                # If the Rz gate is unsupported then the rebase should be skipped
+                # This prevents an error when compiling to the stabilizer backend
+                # where no TK1 replacement can be found for the rebase.
                 passlist.append(self.rebase_pass())
         elif optimisation_level == 1:
             passlist.append(SynthesiseTket())
@@ -433,7 +444,8 @@ class IBMQBackend(Backend):
                     SynthesiseTket(),
                 ]
             )
-        if self._standard_gateset:
+
+        if self._supports_rz:
             passlist.extend([self.rebase_pass(), RemoveRedundancies()])
         if optimisation_level > 0:
             passlist.append(
@@ -447,9 +459,10 @@ class IBMQBackend(Backend):
         return (str, int, int, str)
 
     def rebase_pass(self) -> BasePass:
-        return auto_rebase_pass(
-            {OpType.CX, OpType.X, OpType.SX, OpType.Rz},
-        )
+        if self._primitive_gates == {OpType.X, OpType.SX, OpType.Rz, OpType.ECR}:
+            return ecr_rebase
+        else:
+            return auto_rebase_pass(self._primitive_gates)
 
     def process_circuits(
         self,