+++ tests/backend_test.py
@@ -22,12 +22,13 @@ import numpy as np
 
 import pytest
 
-from qiskit import IBMQ  # type: ignore
-from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
+from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister  # type: ignore
 from qiskit.circuit import Parameter  # type: ignore
 from qiskit.providers.aer.noise.noise_model import NoiseModel  # type: ignore
 from qiskit.providers.aer.noise import ReadoutError  # type: ignore
 from qiskit.providers.aer.noise.errors import depolarizing_error, pauli_error  # type: ignore
+
+from qiskit_ibm_provider import IBMProvider  # type: ignore
 
 from pytket.circuit import Circuit, OpType, BasisOrder, Qubit, reg_eq, Unitary2qBox  # type: ignore
 from pytket.passes import CliffordSimp  # type: ignore
@@ -51,7 +52,10 @@ from pytket.extensions.qiskit import (
     AerUnitaryBackend,
     IBMQEmulatorBackend,
 )
-from pytket.extensions.qiskit import qiskit_to_tk, process_characterisation
+from pytket.extensions.qiskit import (
+    qiskit_to_tk,
+    process_characterisation,
+)
 from pytket.utils.expectations import (
     get_pauli_expectation_value,
     get_operator_expectation_value,
@@ -150,14 +154,9 @@ def test_measures() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_noise() -> None:
-    if not IBMQ.active_account():
-        IBMQ.load_account()
+def test_noise(manila_backend: IBMQBackend) -> None:
 
-    provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_manila")
-
-    noise_model = NoiseModel.from_backend(back)
+    noise_model = NoiseModel.from_backend(manila_backend._backend)
     n_qbs = 5
     c = Circuit(n_qbs, n_qbs)
     x_qbs = [2, 0, 4]
@@ -196,15 +195,11 @@ def test_noise() -> None:
     assert shots.shape == (10, 4)
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_process_characterisation() -> None:
-    if not IBMQ.active_account():
-        IBMQ.load_account()
+def test_process_characterisation(manila_backend: IBMQBackend) -> None:
 
-    provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_manila")
-
-    char = process_characterisation(back)
+    char = process_characterisation(manila_backend._backend)
     arch: Architecture = char.get("Architecture", Architecture([]))
     node_errors: dict = char.get("NodeErrors", {})
     link_errors: dict = char.get("EdgeErrors", {})
@@ -451,6 +446,7 @@ def test_nshots_batching(manila_backend: IBMQBackend) -> None:
         backend._MACHINE_DEBUG = False
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_nshots(manila_emulator_backend: IBMQEmulatorBackend) -> None:
     for b in [AerBackend(), manila_emulator_backend]:
@@ -502,14 +498,9 @@ def test_default_pass(manila_backend: IBMQBackend) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_aer_default_pass() -> None:
-    if not IBMQ.active_account():
-        IBMQ.load_account()
+def test_aer_default_pass(manila_backend: IBMQBackend) -> None:
 
-    provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_manila")
-
-    noise_model = NoiseModel.from_backend(back)
+    noise_model = NoiseModel.from_backend(manila_backend._backend)
     for nm in [None, noise_model]:
         b = AerBackend(nm)
         for ol in range(3):
@@ -681,7 +672,7 @@ def test_operator() -> None:
 
 
 # TKET-1432 this was either too slow or consumed too much memory when bugged
-@pytest.mark.timeout(10)
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 def test_expectation_bug() -> None:
     backend = AerStateBackend()
     # backend.compile_circuit(circuit)
@@ -772,7 +763,7 @@ def test_mixed_circuit() -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_aer_placed_expectation() -> None:
+def test_aer_placed_expectation(manila_backend: IBMQBackend) -> None:
     # bug TKET-695
     n_qbs = 3
     c = Circuit(n_qbs, n_qbs)
@@ -790,13 +781,7 @@ def test_aer_placed_expectation() -> None:
     )
     assert b.get_operator_expectation_value(c, operator) == (-0.5 + 0j)
 
-    if not IBMQ.active_account():
-        IBMQ.load_account()
-
-    provider = IBMQ.providers(hub="ibm-q", group="open")[0]
-    back = provider.get_backend("ibmq_manila")
-
-    noise_model = NoiseModel.from_backend(back)
+    noise_model = NoiseModel.from_backend(manila_backend._backend)
 
     noise_b = AerBackend(noise_model)
 
@@ -824,6 +809,7 @@ def test_operator_expectation_value() -> None:
     assert np.isclose(e, 1.0)
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_ibmq_emulator(manila_emulator_backend: IBMQEmulatorBackend) -> None:
     assert manila_emulator_backend._noise_model is not None
@@ -1101,6 +1087,7 @@ def test_postprocess(lima_backend: IBMQBackend) -> None:
     b.cancel(h)
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_postprocess_emu(manila_emulator_backend: IBMQEmulatorBackend) -> None:
     assert manila_emulator_backend.supports_contextual_optimisation
@@ -1117,7 +1104,7 @@ def test_postprocess_emu(manila_emulator_backend: IBMQEmulatorBackend) -> None:
     assert sum(counts.values()) == 10
 
 
-@pytest.mark.timeout(None)
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_cloud_stabiliser(simulator_stabilizer_backend: IBMQBackend) -> None:
     c = Circuit(2, 2)
@@ -1132,21 +1119,21 @@ def test_cloud_stabiliser(simulator_stabilizer_backend: IBMQBackend) -> None:
 
 
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_available_devices() -> None:
-    backend_info_list = IBMQBackend.available_devices(
-        hub="ibm-q", group="open", project="main"
-    )
+def test_available_devices(ibm_provider: IBMProvider) -> None:
+    backend_info_list = IBMQBackend.available_devices(instance="ibm-q/open/main")
     assert len(backend_info_list) > 0
 
-    provider = IBMQ.providers(hub="ibm-q", group="open")[0]
+    # Check consistency with pytket-qiskit and qiskit provider
+    assert len(backend_info_list) == len(ibm_provider.backends())
 
-    backend_info_list = IBMQBackend.available_devices(account_provider=provider)
+    backend_info_list = IBMQBackend.available_devices(provider=ibm_provider)
     assert len(backend_info_list) > 0
 
     backend_info_list = IBMQBackend.available_devices()
     assert len(backend_info_list) > 0
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_backendinfo_serialization1(
     manila_emulator_backend: IBMQEmulatorBackend,
@@ -1201,6 +1188,7 @@ def test_sim_qubit_order() -> None:
     assert np.isclose(abs(s[2]), 1.0)
 
 
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
 def test_requrired_predicates(manila_emulator_backend: IBMQEmulatorBackend) -> None:
     # https://github.com/CQCL/pytket-qiskit/issues/93
@@ -1214,3 +1202,31 @@ def test_requrired_predicates(manila_emulator_backend: IBMQEmulatorBackend) -> N
             + "not satisfy MaxNQubitsPredicate(5)"
             in str(errorinfo)
         )
+
+
+@pytest.mark.flaky(reruns=3, reruns_delay=10)
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+def test_ecr_gate_compilation(ibm_sherbrooke_backend: IBMQBackend) -> None:
+    assert ibm_sherbrooke_backend.backend_info.gate_set >= {
+        OpType.X,
+        OpType.SX,
+        OpType.Rz,
+        OpType.ECR,
+    }
+    # circuit for an un-routed GHZ state
+    circ = (
+        Circuit(7)
+        .H(0)
+        .CX(0, 1)
+        .CX(0, 2)
+        .CX(0, 3)
+        .CX(0, 4)
+        .CX(0, 5)
+        .CX(0, 6)
+        .measure_all()
+    )
+    for optimisation_level in range(3):
+        compiled_circ = ibm_sherbrooke_backend.get_compiled_circuit(
+            circ, optimisation_level
+        )
+        assert ibm_sherbrooke_backend.valid_circuit(compiled_circ)