+++ docs/changelog.rst
@@ -1,11 +1,25 @@
 Changelog
 ~~~~~~~~~
 
+0.40.0 (June 2023)
+------------------
+
+* IBM devices are now accessed using the `qiskit-ibm-provider <https://github.com/Qiskit/qiskit-ibm-provider>`_ instead of the deprecated :py:class:`IBMQ`. This allows the newest IBM devices and simulators to be accessed through ``pytket-qiskit``. See the updated documentation on `credentials <https://cqcl.github.io/pytket-qiskit/api/index.html#access-and-credentials>`_.
+* The parameters ``hub``, ``group`` and ``project`` are no longer handled as separate arguments in :py:class:`IBMQBackend` and :py:meth:`IBMQBackend.available_devices`. Use ``"instance=f"{hub}/{group}/{project}"`` instead.
+* Added support for the {X, SX, Rz, ECR} in the default compilation pass for :py:class:`IBMQBackend` and :py:class:`IBMQEmulatorBackend`. This is the set of gates used by some of the new IBM devices.
+* Fix to the :py:meth:`tk_to_qiskit` converter to prevent cancellation of redundant gates when converting to qiskit.
+* Handle qiskit circuits with :py:class:`Initialize` and :py:class:`StatePreparation` instructions in the :py:meth:`qiskit_to_tk` converter. The :py:meth:`tk_to_qiskit` converter now handles :py:class:`StatePreparationBox`.
+* Fix handling of control state in :py:meth:`qiskit_to_tk`.
+* Update qiskit version to 0.43.1
+* Update qiskit-ibm-runtime version to 0.11.1
+* Update qiskit-ibm-provider version to 0.6.1
+* Update pytket version to 1.16
+
 0.39.0 (May 2023)
 -----------------
 
 * Updated pytket version requirement to 1.15.
-* The get_compiled_circuit method now allows for optional arguments to override the default settings in the NoiseAwarePlacement
+* The :py:meth:`IBMQBackend.get_compiled_circuit` method now allows for optional arguments to override the default settings in the :py:class:`NoiseAwarePlacement`.
 
 0.38.0 (April 2023)
 -------------------